test(db): split mig-125 delimiter-guard tests into a sibling file (#951)
code-reviewer: rpc-vfr-rt-constraint-regression.integration.test.ts grew to 869
lines (74% over the 500-line test ceiling; standing memory watch said split on
next growth). Move the mig-125 dialog_fill delimiter-guard describe block into
rpc-vfr-rt-mig125-delimiter-guard.integration.test.ts (3 tests); the original
drops back to 613 lines with its other 3 describe blocks intact.

Also restructures the moved afterAll per code-style §7 multi-step cleanup:
- finally { insertedQuestionIds.length = 0 } on the soft-delete step;
- cleanupTestData / cleanupReferenceData in separate try/catch, each gated on
  errors.length === 0 (FK-dependency-correct);
- restored the if (orgId) guard on cleanupTestData (matches the 3 siblings).

No test logic, assertions, or titles changed. Type-check clean; both files green.
